Reorganizing src
