+++ src/App.js
@@ -1,11 +1,11 @@
 import React, { Component, Fragment } from 'react';
 
 // CSS
-import './App.css';
+import './stylesheets/App.css';
 
 // Components
-import Header from './Header'
-import Form from './Form'
+import Header from './components/Header'
+import Form from './components/Form'
 
 class App extends Component {
   render() {
