+++ src/tests/App.test.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+
+// Components
+import App from '../App';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
